[test] apply the new test style to databend-query - optimizers (61/177)
