+++ query/src/optimizers/mod.rs
@@ -12,22 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#[cfg(test)]
-mod optimizer_constant_folding_test;
-#[cfg(test)]
-mod optimizer_expression_transform_test;
-#[cfg(test)]
-mod optimizer_scatters_test;
-#[cfg(test)]
-mod optimizer_statistics_exact_test;
-#[cfg(test)]
-mod optimizer_test;
-#[cfg(test)]
-mod utils_test;
-
-#[cfg(test)]
-mod optimizer_top_n_push_down_test;
-
 mod metrics;
 mod optimizer;
 mod optimizer_constant_folding;
