+++ query/tests/it/optimizers/optimizer_constant_folding.rs
@@ -12,21 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#[cfg(test)]
-mod tests {
-    use common_exception::Result;
+use common_exception::Result;
+use databend_query::optimizers::*;
 
-    use crate::optimizers::*;
+#[test]
+fn test_constant_folding_optimizer() -> Result<()> {
+    struct Test {
+        name: &'static str,
+        query: &'static str,
+        expect: &'static str,
+    }
 
-    #[test]
-    fn test_constant_folding_optimizer() -> Result<()> {
-        struct Test {
-            name: &'static str,
-            query: &'static str,
-            expect: &'static str,
-        }
-
-        let tests: Vec<Test> = vec![
+    let tests: Vec<Test> = vec![
             Test {
                 name: "Projection const recursion",
                 query: "SELECT 1 + 2 + 3",
@@ -173,15 +170,14 @@ mod tests {
             },
         ];
 
-        for test in tests {
-            let ctx = crate::tests::create_query_context()?;
+    for test in tests {
+        let ctx = crate::tests::create_query_context()?;
 
-            let plan = crate::tests::parse_query(test.query, &ctx)?;
-            let mut optimizer = ConstantFoldingOptimizer::create(ctx);
-            let optimized = optimizer.optimize(&plan)?;
-            let actual = format!("{:?}", optimized);
-            assert_eq!(test.expect, actual, "{:#?}", test.name);
-        }
-        Ok(())
+        let plan = crate::tests::parse_query(test.query, &ctx)?;
+        let mut optimizer = ConstantFoldingOptimizer::create(ctx);
+        let optimized = optimizer.optimize(&plan)?;
+        let actual = format!("{:?}", optimized);
+        assert_eq!(test.expect, actual, "{:#?}", test.name);
     }
+    Ok(())
 }
